requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.33.13
botocore==1.33.13
certifi
charset-normalizer==3.3.2
click==8.1.7
einops==0.6.1
exceptiongroup==1.2.2
fastapi==0.103.2
flit_core
greenlet==3.0.3
h11==0.14.0
httptools==0.6.0
idna==3.7
imageio==2.31.2
importlib-metadata==1.7.0
jmespath==1.0.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
natsort==8.4.0
networkx==2.6.3
numpy
opencv-python==4.10.0.84
packaging==24.0
Pillow==9.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.2
scikit-image==0.19.3
scipy==1.7.3
six
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.27.0
tifffile==2021.11.2
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==1.26.19
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
